+++ cx-portal/src/components/shared/basic/ReleaseProcess/AppPage/index.tsx
@@ -25,6 +25,7 @@ import {
   UploadStatus,
   DropArea,
   DropAreaProps,
+  PageSnackbar,
 } from 'cx-portal-shared-components'
 import { useTranslation } from 'react-i18next'
 import { Divider, InputLabel } from '@mui/material'
@@ -42,6 +43,7 @@ import {
 } from 'features/appManagement/slice'
 import {
   DocumentTypeId,
+  useDeleteAppReleaseDocumentMutation,
   useFetchAppStatusQuery,
   useUpdateappMutation,
   useUpdateDocumentUploadMutation,
@@ -68,18 +70,26 @@ type FormDataType = {
 
 export default function AppPage() {
   const { t } = useTranslation()
+  const dispatch = useDispatch()
+  const appId = useSelector(appIdSelector)
+
   const [appPageNotification, setAppPageNotification] = useState(false)
   const [appPageSnackbar, setAppPageSnackbar] = useState<boolean>(false)
-  const dispatch = useDispatch()
+  const [deleteSuccess, setDeleteSuccess] = useState(false)
   const [updateapp] = useUpdateappMutation()
   const [updateDocumentUpload] = useUpdateDocumentUploadMutation()
-  const appId = useSelector(appIdSelector)
   const longDescriptionMaxLength = 2000
   const fetchAppStatus = useFetchAppStatusQuery(appId ?? '', {
     refetchOnMountOrArgChange: true,
   }).data
   const appStatusData: any = useSelector(appStatusDataSelector)
   const statusData = fetchAppStatus || appStatusData
+
+  const [deleteAppReleaseDocument, deleteResponse] =
+    useDeleteAppReleaseDocumentMutation()
+  useEffect(() => {
+    deleteResponse.isSuccess && setDeleteSuccess(true)
+  }, [deleteResponse])
 
   const defaultValues = {
     longDescriptionEN:
@@ -126,17 +136,21 @@ export default function AppPage() {
   const defaultuploadAppContract = defaultValues.uploadAppContract
 
   useEffect(() => {
-    const images = defaultImages?.map((item: { documentName: string }) => {
-      return {
-        name: item.documentName,
-        status: UploadStatus.UPLOAD_SUCCESS,
+    const images = defaultImages?.map(
+      (item: { documentId: string; documentName: string }) => {
+        return {
+          id: item.documentId,
+          name: item.documentName,
+          status: UploadStatus.UPLOAD_SUCCESS,
+        }
       }
-    })
+    )
 
     if (images.length > 0) {
       const setFileStatus = (fileIndex: number, status: UploadFileStatus) => {
         const nextFiles = images
         nextFiles[fileIndex] = {
+          id: images[fileIndex].id,
           name: images[fileIndex].name,
           status,
         }
@@ -153,6 +167,9 @@ export default function AppPage() {
       Object.keys(defaultuploadDataPrerequisits).length > 0
     ) {
       setValue('uploadDataPrerequisits', {
+        id:
+          defaultuploadDataPrerequisits &&
+          defaultuploadDataPrerequisits[0]?.documentId,
         name:
           defaultuploadDataPrerequisits &&
           defaultuploadDataPrerequisits[0]?.documentName,
@@ -166,6 +183,9 @@ export default function AppPage() {
       Object.keys(defaultuploadTechnicalGuide).length > 0
     ) {
       setValue('uploadTechnicalGuide', {
+        id:
+          defaultuploadTechnicalGuide &&
+          defaultuploadTechnicalGuide[0]?.documentId,
         name:
           defaultuploadTechnicalGuide &&
           defaultuploadTechnicalGuide[0]?.documentName,
@@ -179,6 +199,7 @@ export default function AppPage() {
       Object.keys(defaultuploadAppContract).length > 0
     ) {
       setValue('uploadAppContract', {
+        id: defaultuploadAppContract && defaultuploadAppContract[0]?.documentId,
         name:
           defaultuploadAppContract && defaultuploadAppContract[0]?.documentName,
         status: UploadStatus.UPLOAD_SUCCESS,
@@ -200,6 +221,7 @@ export default function AppPage() {
     const value = getValues(fieldName)
 
     setValue(fieldName, {
+      id: value.id,
       name: value.name,
       size: value.size,
       status,
@@ -263,6 +285,7 @@ export default function AppPage() {
       const setFileStatus = (fileIndex: number, status: UploadFileStatus) => {
         const nextFiles = [...getValues().images] as any[]
         nextFiles[fileIndex] = {
+          id: value[fileIndex].id,
           name: value[fileIndex].name,
           size: value[fileIndex].size,
           status,
@@ -444,6 +467,10 @@ export default function AppPage() {
                   maxFilesToUpload={3}
                   maxFileSize={819200}
                   DropArea={renderDropArea}
+                  handleDelete={(documentId: string) => {
+                    setDeleteSuccess(false)
+                    documentId && deleteAppReleaseDocument(documentId)
+                  }}
                 />
               )
             }}
@@ -485,6 +512,10 @@ export default function AppPage() {
                   maxFilesToUpload: 1,
                   maxFileSize: 819200,
                   size: 'small',
+                }}
+                handleDelete={(documentId: string) => {
+                  setDeleteSuccess(false)
+                  documentId && deleteAppReleaseDocument(documentId)
                 }}
               />
               {item === 'uploadDataPrerequisits' &&
@@ -578,6 +609,15 @@ export default function AppPage() {
           onAppPageSubmit(data, 'saveAndProceed')
         )}
         isValid={isValid}
+      />
+      <PageSnackbar
+        autoClose
+        description={t(
+          'content.apprelease.contractAndConsent.documentDeleteSuccess'
+        )}
+        open={deleteSuccess}
+        severity={'success'}
+        showIcon
       />
     </div>
   )