+++ cx-portal/src/components/shared/basic/ReleaseProcess/components/ConnectorFormInputField.tsx
@@ -50,6 +50,7 @@ export const ConnectorFormInputField = ({
   defaultValues,
   enableDeleteIcon,
   handleDownload,
+  handleDelete,
   size,
 }: any) => {
   const renderDropArea = (props: DropAreaProps, size: any) => {
@@ -87,12 +88,13 @@ export const ConnectorFormInputField = ({
         } else if (type === 'dropzone') {
           return (
             <Dropzone
-              files={value ? [value] : undefined}
+              files={value && value.name ? [value] : undefined}
               onChange={([file]) => {
                 trigger(name)
                 onChange(file)
               }}
               handleDownload={handleDownload}
+              handleDelete={handleDelete}
               acceptFormat={acceptFormat}
               maxFilesToUpload={maxFilesToUpload}
               maxFileSize={maxFileSize}