+++ cx-portal/src/features/appManagement/apiSlice.ts
@@ -197,6 +197,7 @@ export enum DocumentTypeId {
 export const apiSlice = createApi({
   reducerPath: 'rtk/appManagement',
   baseQuery: fetchBaseQuery(apiBaseQuery()),
+  tagTypes: ['App'],
   endpoints: (builder) => ({
     fetchUseCases: builder.query<useCasesItem[], void>({
       query: () => `/api/administration/staticdata/usecases`,
@@ -249,6 +250,7 @@ export const apiSlice = createApi({
     }),
     fetchAppStatus: builder.query<AppStatusDataState, string>({
       query: (appId) => `api/apps/appreleaseprocess/${appId}/appStatus`,
+      providesTags: ['App'],
     }),
     fetchAgreementData: builder.query<AgreementType[], void>({
       query: () => `api/apps/AppReleaseProcess/agreementData`,
@@ -289,6 +291,7 @@ export const apiSlice = createApi({
         url: `/api/apps/appreleaseprocess/documents/${documentId}`,
         method: 'DELETE',
       }),
+      invalidatesTags: ['App'],
     }),
     fetchRolesData: builder.query<rolesType[], string>({
       query: (appId: string) => `api/administration/user/app/${appId}/roles`,